Revert "refactor : model 패키지로 이동"
This reverts commit af70b74715a38aac1f41a883b5b15a841774d02c.
